+++ addons/addons.go
@@ -18,6 +18,7 @@ import (
 	_ "github.com/invopop/gobl/addons/es/tbai"
 	_ "github.com/invopop/gobl/addons/es/verifactu"
 	_ "github.com/invopop/gobl/addons/eu/en16931"
+	_ "github.com/invopop/gobl/addons/fi/finvoice"
 	_ "github.com/invopop/gobl/addons/fr/choruspro"
 	_ "github.com/invopop/gobl/addons/fr/facturx"
 	_ "github.com/invopop/gobl/addons/gr/mydata"